Created overworld UI class
Renamed ashleys ui class into the battleUI and added overworld ui ive
been working on for the last few weeks. Got rid of unnessesarry empty
lines in ashleys ui class. There are aspects of each class in eachother
that need to be moved such as screen scroll needs to be in battle not
overworld.